Replace particles layer with flow geometries
